gameplayer/requirements.txt
pyautogui==0.9.54
keyboard==0.13.5
python-dotenv==1.0.0
pynput==1.7.6
